backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
python-multipart==0.0.6
pillow==10.0.0
pytesseract==0.3.10
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.3
psycopg2-binary==2.9.9

# Note: Tesseract-OCR binary must be installed separately on the host OS.